Refactor network and element classes
