+++ roseau/load_flow/models/branches.py
@@ -1,5 +1,6 @@
 import logging
 import warnings
+from abc import abstractmethod
 from functools import cached_property
 from typing import TYPE_CHECKING, Self
 
@@ -36,6 +37,7 @@ class AbstractBranch(Element[_CyB_co]):
         :doc:`Switch model documentation </models/Switch>`
     """
 
+    @abstractmethod
     def __init__(
         self, id: Id, bus1: Bus, bus2: Bus, *, phases1: str, phases2: str, geometry: BaseGeometry | None
     ) -> None:
@@ -60,8 +62,6 @@ class AbstractBranch(Element[_CyB_co]):
             geometry:
                 The geometry of the branch.
         """
-        if type(self) is AbstractBranch:
-            raise TypeError("Can't instantiate abstract class AbstractBranch")
         super().__init__(id)
         self._check_phases(id=id, phases1=phases1)
         self._check_phases(id=id, phases2=phases2)
