+++ roseau/load_flow/models/potential_refs.py
@@ -25,7 +25,7 @@ class PotentialRef(Element[CyPotentialRef | CyDeltaPotentialRef]):
     bus, the sum of the potentials of the specified phases is set to 0V.
     """
 
-    element_type: Final = "potential reference"
+    element_type: Final = "potential ref"
     allowed_phases: Final = frozenset({"a", "b", "c", "n"} | Bus.allowed_phases)
 
     @deprecate_renamed_parameter(old_name="phase", new_name="phases", version="0.10.0", category=DeprecationWarning)
