+++ roseau/load_flow/utils/__init__.py
@@ -18,9 +18,15 @@ from roseau.load_flow.utils.dtypes import (
     VoltagePhaseDtype,
 )
 from roseau.load_flow.utils.exceptions import find_stack_level
-from roseau.load_flow.utils.helpers import CaseInsensitiveStrEnum, count_repr, id_sort_key, one_or_more_repr
+from roseau.load_flow.utils.helpers import (
+    CaseInsensitiveStrEnum,
+    abstractattrs,
+    count_repr,
+    id_sort_key,
+    one_or_more_repr,
+)
 from roseau.load_flow.utils.log import set_logging_config
-from roseau.load_flow.utils.mixins import CatalogueMixin, Identifiable, JsonMixin
+from roseau.load_flow.utils.mixins import AbstractElement, AbstractNetwork, CatalogueMixin, Identifiable, JsonMixin
 from roseau.load_flow.utils.versions import show_versions
 
 __all__ = [
@@ -28,6 +34,8 @@ __all__ = [
     "Identifiable",
     "JsonMixin",
     "CatalogueMixin",
+    "AbstractElement",
+    "AbstractNetwork",
     # Exceptions and warnings
     "find_stack_level",
     "deprecate_nonkeyword_arguments",
@@ -52,6 +60,8 @@ __all__ = [
     "id_sort_key",
     # Enums
     "CaseInsensitiveStrEnum",
+    # Decorators
+    "abstractattrs",
 ]
