+++ roseau/load_flow_single/models/buses.py
@@ -105,7 +105,7 @@ class Bus(AbstractTerminal[CyBus]):
         self._invalidate_network_results()
         self._initialized = True
         self._initialized_by_the_user = True
-        if self._cy_element is not None:
+        if self._cy_initialized:
             self._cy_element.initialize_potentials(np.array([self._initial_voltage / SQRT3, 0], dtype=np.complex128))
 
     @property
