+++ roseau/load_flow_single/tests/test_electrical_network.py
@@ -120,7 +120,7 @@ def test_connect_and_disconnect():
     with pytest.raises(RoseauLoadFlowException) as e:
         en._disconnect_element(line)
     assert e.value.msg == (
-        "<Line: id='line', bus1='source', bus2='load bus'> is a Line and it cannot be disconnected from a network."
+        "<Line: id='line', bus1='source', bus2='load bus'> is a line and cannot be disconnected from a network."
     )
     assert e.value.code == RoseauLoadFlowExceptionCode.BAD_ELEMENT_OBJECT
